+++ R/rds.R
@@ -5,6 +5,7 @@
 #' than time.
 #'
 #' @param file The file path to read from/write to.
+#' @param refhook A function to handle reference objects.
 #' @keywords internal
 #' @export
 #' @examples
@@ -14,11 +15,11 @@
 #' \dontrun{
 #' write_rds(mtcars, "compressed_mtc.rds", "xz", compression = 9L)
 #' }
-read_rds <- function(file) {
+read_rds <- function(file, refhook = NULL) {
   con <- file(file)
   on.exit(close(con))
 
-  readRDS(con)
+  readRDS(con, refhook = refhook)
 }
 
 
@@ -31,11 +32,12 @@ read_rds <- function(file) {
 #'   the space-time trade-off of different compression methods with
 #'   `compression`. See [connections()] for more details.
 #' @param path \Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
+#' @inheritParams read_rds
 #' @return `write_rds()` returns `x`, invisibly.
 #' @rdname read_rds
 #' @export
 write_rds <- function(x, file, compress = c("none", "gz", "bz2", "xz"),
-                      version = 2, path = deprecated(), ...) {
+                      version = 2, refhook = NULL, path = deprecated(), ...) {
   if (is_present(path)) {
     deprecate_warn("1.4.0", "write_rds(path = )", "write_rds(file = )")
     file <- path
@@ -49,7 +51,7 @@ write_rds <- function(x, file, compress = c("none", "gz", "bz2", "xz"),
     xz = xzfile(file, ...)
   )
   on.exit(close(con), add = TRUE)
-  saveRDS(x, con, version = version)
+  saveRDS(x, con, version = version, refhook = refhook)
 
   invisible(x)
 }